* This is a major reorganization to use Notes and Coordinator objects.
I've also moved the class names around. Instead of combining
components in a single class, I'll just make them separate. This
makes it easier for me to remember where thigns are.

I've added a new reporter: MyCPAN::App::DPAN::Reporter::Minimal.
This does away with all of the YAML, which was a real drain on
time.