* Add X-SVN header to 02packages, dump run info in error output
If the Minimal reporter sees a .svn directory in the repo, it
adds an 'X-SVN' header to 02packages.details.txt.gz. The value
is merely '$Id$' so you can see which version of DPAN you have
if you check it out from the server. This is handy when you
intend to use older versions for testing, etc.

The Reporter dumps into the error report everything it knows
so you can inspect the state of the indexer to guess what
might have gone wrong.

When dpan is done, it now prints a short message reminding
people to reload their CPAN indexes.