* Add support for a cleanup class
The configuration directive cleanup_class takes the name of a
class that dpan will execute after it has finished all of its
work. It executes the run() method in that class, passing it
the application object.

It's your class and you can do anything you like. It's your
chance to insert extra steps into the dpan process.